requirements.txt
aiofiles==0.5.0
aiohttp==3.7.4.post0
aniso8601==7.0.0
APScheduler==3.7.0
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
coverage==6.4
cryptography==3.4.8
cssselect==1.1.0
cssutils==2.3.0
databases==0.4.3
dnspython==2.1.0
docopt==0.6.2
ecdsa==0.17.0
email-validator==1.1.2
fastapi==0.65.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.7.1
mailjet-rest==1.3.4
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.22.4
packaging==21.3
pandas==1.2.4
passlib==1.7.4
pika==1.2.0
pipreqs==0.4.11
pluggy==1.0.0
premailer==3.10.0
promise==2.3
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==3.0.9
pyshorteners==1.0.1
pytest==7.1.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
tomli==2.0.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
XlsxWriter==3.0.2
yarg==0.1.9
yarl==1.7.2
